MediTrakCore files moved
