+++ app/src/main/cpp/MediTrakCore/CMakeLists.txt
@@ -1,0 +1,7 @@
+include_directories(DatabaseController Dose Medication)
+
+add_subdirectory(DatabaseController)
+add_subdirectory(Dose)
+add_subdirectory(Medication)
+
+target_link_libraries(DatabaseController Dose Medication)
